Move theme system out of core to a client-side ThemeProvider
Themes are purely a client/rendering concern; the simulation never reads
colors. Relocate Theme/PastelTheme/PastelThemeDark/Colors/ColorAllocator
from src/core/configuration to src/client/theme and remove Config.theme().
Consumers now read the active theme from a themeProvider singleton, which
resets its color allocators per game to match the previous per-Config
lifecycle. No behavior change.
