+++ src/core/configuration/Config.ts
@@ -21,9 +21,6 @@ import { UserSettings } from "../game/UserSettings";
 import { GameConfig, TeamCountConfig } from "../Schemas";
 import { NukeType } from "../StatsSchemas";
 import { assertNever, sigmoid, toInt, within } from "../Util";
-import { PastelTheme } from "./PastelTheme";
-import { PastelThemeDark } from "./PastelThemeDark";
-import { Theme } from "./Theme";
 
 declare global {
   interface Window {
@@ -84,8 +81,6 @@ export const JwksSchema = z.object({
 export const SAM_CONSTRUCTION_TICKS = 30 * 10;
 
 export class Config {
-  private pastelTheme: PastelTheme = new PastelTheme();
-  private pastelThemeDark: PastelThemeDark = new PastelThemeDark();
   private unitInfoCache = new Map<UnitType, UnitInfo>();
   constructor(
     private _gameConfig: GameConfig,
@@ -561,11 +556,6 @@ export class Config {
   }
   numBots(): number {
     return this.bots();
-  }
-  theme(): Theme {
-    return this.userSettings()?.darkMode()
-      ? this.pastelThemeDark
-      : this.pastelTheme;
   }
 
   attackLogic(
